pyproject.toml
[project]
name = "test-task"
version = "0.1.0"
description = "test task for backend developer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "celery>=5.6.0",
    "django>=5.2",
    "django-celery-results>=2.6.0",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.29.0",
    "flower>=2.0.1",
    "psycopg2-binary>=2.9.11",
]

[dependency-groups]
dev = [
 "pytest-django>=4.9.0",
 "pre-commit>=4.5.0",
 "pytest-mock>=3.14.0",
 "pytest>=9.0.1",
 "ruff>=0.14.7"
]
prod = ["gunicorn>=23.0.0"]

[tool.ruff]
line-length = 120
target-version = "py310"
extend-exclude = ["*/migrations/*"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
docstring-code-format = false

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["tests"]
addopts = "--tb=short --reuse-db"
